DRTII-1848 Reinstate call to updatePrediction
